+++ src/platform_impl/linux/wayland/env.rs
@@ -13,6 +13,7 @@ use sctk::reexports::protocols::xdg_shell::client::xdg_wm_base::XdgWmBase;
 use sctk::reexports::protocols::unstable::relative_pointer::v1::client::zwp_relative_pointer_manager_v1::ZwpRelativePointerManagerV1;
 use sctk::reexports::protocols::unstable::pointer_constraints::v1::client::zwp_pointer_constraints_v1::ZwpPointerConstraintsV1;
 use sctk::reexports::protocols::unstable::text_input::v3::client::zwp_text_input_manager_v3::ZwpTextInputManagerV3;
+use sctk::reexports::protocols::staging::xdg_activation::v1::client::xdg_activation_v1::XdgActivationV1;
 
 use sctk::environment::{Environment, SimpleGlobal};
 use sctk::output::{OutputHandler, OutputHandling, OutputInfo, OutputStatusListener};
@@ -24,17 +25,26 @@ use sctk::shm::ShmHandler;
 #[derive(Debug, Clone, Copy)]
 pub struct WindowingFeatures {
     cursor_grab: bool,
+    xdg_activation: bool,
 }
 
 impl WindowingFeatures {
     /// Create `WindowingFeatures` based on the presented interfaces.
     pub fn new(env: &Environment<WinitEnv>) -> Self {
         let cursor_grab = env.get_global::<ZwpPointerConstraintsV1>().is_some();
-        Self { cursor_grab }
+        let xdg_activation = env.get_global::<XdgActivationV1>().is_some();
+        Self {
+            cursor_grab,
+            xdg_activation,
+        }
     }
 
     pub fn cursor_grab(&self) -> bool {
         self.cursor_grab
+    }
+
+    pub fn xdg_activation(&self) -> bool {
+        self.xdg_activation
     }
 }
 
@@ -50,6 +60,7 @@ sctk::environment!(WinitEnv,
         ZwpRelativePointerManagerV1 => relative_pointer_manager,
         ZwpPointerConstraintsV1 => pointer_constraints,
         ZwpTextInputManagerV3 => text_input_manager,
+        XdgActivationV1 => xdg_activation,
     ],
     multis = [
         WlSeat => seats,
@@ -78,6 +89,8 @@ pub struct WinitEnv {
     text_input_manager: SimpleGlobal<ZwpTextInputManagerV3>,
 
     decoration_manager: SimpleGlobal<ZxdgDecorationManagerV1>,
+
+    xdg_activation: SimpleGlobal<XdgActivationV1>,
 }
 
 impl WinitEnv {
@@ -109,6 +122,9 @@ impl WinitEnv {
         // IME handling.
         let text_input_manager = SimpleGlobal::new();
 
+        // Surface activation.
+        let xdg_activation = SimpleGlobal::new();
+
         Self {
             seats,
             outputs,
@@ -120,6 +136,7 @@ impl WinitEnv {
             relative_pointer_manager,
             pointer_constraints,
             text_input_manager,
+            xdg_activation,
         }
     }
 }