+++ src/platform_impl/linux/wayland/window/mod.rs
@@ -7,21 +7,17 @@ use sctk::reexports::client::Display;
 
 use sctk::reexports::calloop;
 
-use sctk::window::{
-    ARGBColor, ButtonColorSpec, ColorSpec, ConceptConfig, ConceptFrame, Decorations,
-};
-
 use raw_window_handle::unix::WaylandHandle;
+use sctk::window::{Decorations, FallbackFrame};
 
 use crate::dpi::{LogicalSize, PhysicalPosition, PhysicalSize, Position, Size};
 use crate::error::{ExternalError, NotSupportedError, OsError as RootOsError};
 use crate::monitor::MonitorHandle as RootMonitorHandle;
-use crate::platform::unix::{ARGBColor as LocalARGBColor, Button, ButtonState, Element, Theme};
 use crate::platform_impl::{
     MonitorHandle as PlatformMonitorHandle, OsError,
     PlatformSpecificWindowBuilderAttributes as PlatformAttributes,
 };
-use crate::window::{CursorIcon, Fullscreen, WindowAttributes};
+use crate::window::{CursorIcon, Fullscreen, UserAttentionType, WindowAttributes};
 
 use super::env::WindowingFeatures;
 use super::event_loop::WinitState;
@@ -103,7 +99,7 @@ impl Window {
         let theme_manager = event_loop_window_target.theme_manager.clone();
         let mut window = event_loop_window_target
             .env
-            .create_window::<ConceptFrame, _>(
+            .create_window::<FallbackFrame, _>(
                 surface.clone(),
                 Some(theme_manager),
                 (width, height),
@@ -201,7 +197,12 @@ impl Window {
         let window_requests = Arc::new(Mutex::new(Vec::with_capacity(64)));
 
         // Create a handle that performs all the requests on underlying sctk a window.
-        let window_handle = WindowHandle::new(window, size.clone(), window_requests.clone());
+        let window_handle = WindowHandle::new(
+            &event_loop_window_target.env,
+            window,
+            size.clone(),
+            window_requests.clone(),
+        );
 
         let mut winit_state = event_loop_window_target.state.borrow_mut();
 
@@ -213,17 +214,14 @@ impl Window {
 
         let windowing_features = event_loop_window_target.windowing_features;
 
-        // Send all updates to the server.
-        let wayland_source = &event_loop_window_target.wayland_source;
-        let event_loop_handle = &event_loop_window_target.event_loop_handle;
-
         // To make our window usable for drawing right away we must `ack` a `configure`
         // from the server, the acking part here is done by SCTK window frame, so we just
         // need to sync with server so it'll be done automatically for us.
-        event_loop_handle.with_source(&wayland_source, |event_queue| {
-            let event_queue = event_queue.queue();
+        {
+            let mut wayland_source = event_loop_window_target.wayland_dispatcher.as_source_mut();
+            let event_queue = wayland_source.queue();
             let _ = event_queue.sync_roundtrip(&mut *winit_state, |_, _, _| unreachable!());
-        });
+        }
 
         // We all praise GNOME for these 3 lines of pure magic. If we don't do that,
         // GNOME will shrink our window a bit for the size of the decorations. I guess it
@@ -258,9 +256,7 @@ impl Window {
 
     #[inline]
     pub fn set_title(&self, title: &str) {
-        let title_request = WindowRequest::Title(title.to_owned());
-        self.window_requests.lock().unwrap().push(title_request);
-        self.event_loop_awakener.ping();
+        self.send_request(WindowRequest::Title(title.to_owned()));
     }
 
     #[inline]
@@ -292,9 +288,7 @@ impl Window {
 
     #[inline]
     pub fn request_redraw(&self) {
-        let redraw_request = WindowRequest::Redraw;
-        self.window_requests.lock().unwrap().push(redraw_request);
-        self.event_loop_awakener.ping();
+        self.send_request(WindowRequest::Redraw);
     }
 
     #[inline]
@@ -312,12 +306,7 @@ impl Window {
         let size = size.to_logical::<u32>(scale_factor);
         *self.size.lock().unwrap() = size;
 
-        let frame_size_request = WindowRequest::FrameSize(size);
-        self.window_requests
-            .lock()
-            .unwrap()
-            .push(frame_size_request);
-        self.event_loop_awakener.ping();
+        self.send_request(WindowRequest::FrameSize(size));
     }
 
     #[inline]
@@ -325,9 +314,7 @@ impl Window {
         let scale_factor = self.scale_factor() as f64;
         let size = dimensions.map(|size| size.to_logical::<u32>(scale_factor));
 
-        let min_size_request = WindowRequest::MinSize(size);
-        self.window_requests.lock().unwrap().push(min_size_request);
-        self.event_loop_awakener.ping();
+        self.send_request(WindowRequest::MinSize(size));
     }
 
     #[inline]
@@ -335,19 +322,12 @@ impl Window {
         let scale_factor = self.scale_factor() as f64;
         let size = dimensions.map(|size| size.to_logical::<u32>(scale_factor));
 
-        let max_size_request = WindowRequest::MaxSize(size);
-        self.window_requests.lock().unwrap().push(max_size_request);
-        self.event_loop_awakener.ping();
+        self.send_request(WindowRequest::MaxSize(size));
     }
 
     #[inline]
     pub fn set_resizable(&self, resizable: bool) {
-        let resizeable_request = WindowRequest::Resizeable(resizable);
-        self.window_requests
-            .lock()
-            .unwrap()
-            .push(resizeable_request);
-        self.event_loop_awakener.ping();
+        self.send_request(WindowRequest::Resizeable(resizable));
     }
 
     #[inline]
@@ -359,9 +339,7 @@ impl Window {
 
     #[inline]
     pub fn set_decorations(&self, decorate: bool) {
-        let decorate_request = WindowRequest::Decorate(decorate);
-        self.window_requests.lock().unwrap().push(decorate_request);
-        self.event_loop_awakener.ping();
+        self.send_request(WindowRequest::Decorate(decorate));
     }
 
     #[inline]
@@ -371,9 +349,7 @@ impl Window {
             return;
         }
 
-        let minimize_request = WindowRequest::Minimize;
-        self.window_requests.lock().unwrap().push(minimize_request);
-        self.event_loop_awakener.ping();
+        self.send_request(WindowRequest::Minimize);
     }
 
     #[inline]
@@ -383,9 +359,7 @@ impl Window {
 
     #[inline]
     pub fn set_maximized(&self, maximized: bool) {
-        let maximize_request = WindowRequest::Maximize(maximized);
-        self.window_requests.lock().unwrap().push(maximize_request);
-        self.event_loop_awakener.ping();
+        self.send_request(WindowRequest::Maximize(maximized));
     }
 
     #[inline]
@@ -421,154 +395,17 @@ impl Window {
             None => WindowRequest::UnsetFullscreen,
         };
 
-        self.window_requests
-            .lock()
-            .unwrap()
-            .push(fullscreen_request);
-        self.event_loop_awakener.ping();
-    }
-
-    #[inline]
-    pub fn set_theme<T: Theme>(&self, theme: T) {
-        // First buttons is minimize, then maximize, and then close.
-        let buttons: Vec<(ButtonColorSpec, ButtonColorSpec)> =
-            [Button::Minimize, Button::Maximize, Button::Close]
-                .iter()
-                .map(|button| {
-                    let button = *button;
-                    let idle_active_bg = theme
-                        .button_color(button, ButtonState::Idle, false, true)
-                        .into();
-                    let idle_inactive_bg = theme
-                        .button_color(button, ButtonState::Idle, false, false)
-                        .into();
-                    let idle_active_icon = theme
-                        .button_color(button, ButtonState::Idle, true, true)
-                        .into();
-                    let idle_inactive_icon = theme
-                        .button_color(button, ButtonState::Idle, true, false)
-                        .into();
-                    let idle_bg = ColorSpec {
-                        active: idle_active_bg,
-                        inactive: idle_inactive_bg,
-                    };
-                    let idle_icon = ColorSpec {
-                        active: idle_active_icon,
-                        inactive: idle_inactive_icon,
-                    };
-
-                    let hovered_active_bg = theme
-                        .button_color(button, ButtonState::Hovered, false, true)
-                        .into();
-                    let hovered_inactive_bg = theme
-                        .button_color(button, ButtonState::Hovered, false, false)
-                        .into();
-                    let hovered_active_icon = theme
-                        .button_color(button, ButtonState::Hovered, true, true)
-                        .into();
-                    let hovered_inactive_icon = theme
-                        .button_color(button, ButtonState::Hovered, true, false)
-                        .into();
-                    let hovered_bg = ColorSpec {
-                        active: hovered_active_bg,
-                        inactive: hovered_inactive_bg,
-                    };
-                    let hovered_icon = ColorSpec {
-                        active: hovered_active_icon,
-                        inactive: hovered_inactive_icon,
-                    };
-
-                    let disabled_active_bg = theme
-                        .button_color(button, ButtonState::Disabled, false, true)
-                        .into();
-                    let disabled_inactive_bg = theme
-                        .button_color(button, ButtonState::Disabled, false, false)
-                        .into();
-                    let disabled_active_icon = theme
-                        .button_color(button, ButtonState::Disabled, true, true)
-                        .into();
-                    let disabled_inactive_icon = theme
-                        .button_color(button, ButtonState::Disabled, true, false)
-                        .into();
-                    let disabled_bg = ColorSpec {
-                        active: disabled_active_bg,
-                        inactive: disabled_inactive_bg,
-                    };
-                    let disabled_icon = ColorSpec {
-                        active: disabled_active_icon,
-                        inactive: disabled_inactive_icon,
-                    };
-
-                    let button_bg = ButtonColorSpec {
-                        idle: idle_bg,
-                        hovered: hovered_bg,
-                        disabled: disabled_bg,
-                    };
-                    let button_icon = ButtonColorSpec {
-                        idle: idle_icon,
-                        hovered: hovered_icon,
-                        disabled: disabled_icon,
-                    };
-
-                    (button_icon, button_bg)
-                })
-                .collect();
-
-        let minimize_button = Some(buttons[0]);
-        let maximize_button = Some(buttons[1]);
-        let close_button = Some(buttons[2]);
-
-        // The first color is bar, then separator, and then text color.
-        let titlebar_colors: Vec<ColorSpec> = [Element::Bar, Element::Separator, Element::Text]
-            .iter()
-            .map(|element| {
-                let element = *element;
-                let active = theme.element_color(element, true).into();
-                let inactive = theme.element_color(element, false).into();
-
-                ColorSpec { active, inactive }
-            })
-            .collect();
-
-        let primary_color = titlebar_colors[0];
-        let secondary_color = titlebar_colors[1];
-        let title_color = titlebar_colors[2];
-
-        let title_font = theme.font();
-
-        let concept_config = ConceptConfig {
-            primary_color,
-            secondary_color,
-            title_color,
-            title_font,
-            minimize_button,
-            maximize_button,
-            close_button,
-        };
-
-        let theme_request = WindowRequest::Theme(concept_config);
-        self.window_requests.lock().unwrap().push(theme_request);
-        self.event_loop_awakener.ping();
+        self.send_request(fullscreen_request);
     }
 
     #[inline]
     pub fn set_cursor_icon(&self, cursor: CursorIcon) {
-        let cursor_icon_request = WindowRequest::NewCursorIcon(cursor);
-        self.window_requests
-            .lock()
-            .unwrap()
-            .push(cursor_icon_request);
-        self.event_loop_awakener.ping();
+        self.send_request(WindowRequest::NewCursorIcon(cursor));
     }
 
     #[inline]
     pub fn set_cursor_visible(&self, visible: bool) {
-        let cursor_visible_request = WindowRequest::ShowCursor(visible);
-        self.window_requests
-            .lock()
-            .unwrap()
-            .push(cursor_visible_request);
-        self.event_loop_awakener.ping();
+        self.send_request(WindowRequest::ShowCursor(visible));
     }
 
     #[inline]
@@ -577,14 +414,18 @@ impl Window {
             return Err(ExternalError::NotSupported(NotSupportedError::new()));
         }
 
-        let cursor_grab_request = WindowRequest::GrabCursor(grab);
-        self.window_requests
-            .lock()
-            .unwrap()
-            .push(cursor_grab_request);
-        self.event_loop_awakener.ping();
+        self.send_request(WindowRequest::GrabCursor(grab));
 
         Ok(())
+    }
+
+    pub fn request_user_attention(&self, request_type: Option<UserAttentionType>) {
+        if !self.windowing_features.xdg_activation() {
+            warn!("`request_user_attention` isn't supported");
+            return;
+        }
+
+        self.send_request(WindowRequest::Attention(request_type));
     }
 
     #[inline]
@@ -601,12 +442,7 @@ impl Window {
 
     #[inline]
     pub fn drag_window(&self) -> Result<(), ExternalError> {
-        let drag_window_request = WindowRequest::DragWindow;
-        self.window_requests
-            .lock()
-            .unwrap()
-            .push(drag_window_request);
-        self.event_loop_awakener.ping();
+        self.send_request(WindowRequest::DragWindow);
 
         Ok(())
     }
@@ -615,12 +451,7 @@ impl Window {
     pub fn set_ime_position(&self, position: Position) {
         let scale_factor = self.scale_factor() as f64;
         let position = position.to_logical(scale_factor);
-        let ime_position_request = WindowRequest::IMEPosition(position);
-        self.window_requests
-            .lock()
-            .unwrap()
-            .push(ime_position_request);
-        self.event_loop_awakener.ping();
+        self.send_request(WindowRequest::IMEPosition(position));
     }
 
     #[inline]
@@ -660,22 +491,16 @@ impl Window {
             ..WaylandHandle::empty()
         }
     }
-}
 
-impl From<LocalARGBColor> for ARGBColor {
-    fn from(color: LocalARGBColor) -> Self {
-        let a = color.a;
-        let r = color.r;
-        let g = color.g;
-        let b = color.b;
-        Self { a, r, g, b }
+    #[inline]
+    fn send_request(&self, request: WindowRequest) {
+        self.window_requests.lock().unwrap().push(request);
+        self.event_loop_awakener.ping();
     }
 }
 
 impl Drop for Window {
     fn drop(&mut self) {
-        let close_request = WindowRequest::Close;
-        self.window_requests.lock().unwrap().push(close_request);
-        self.event_loop_awakener.ping();
+        self.send_request(WindowRequest::Close);
     }
 }